requirements.txt
xmpppy==0.5.0rc1